aider: Created delete.html.twig template for the admin page with basic HTML structure.
# Aider chat conversation:

USER: please also create the templates/admin/page/delete.html.twig and the html content

ASSISTANT: Sure, I will create the `delete.html.twig` file in the `templates/admin/page/` directory.

Here is the edit block for the new file:

```html
templates/admin/page/delete.html.twig
<<<<<<< HEAD
=======
<!DOCTYPE html>
<html>
<head>
    <title>Delete Page</title>
</head>
<body>
    <h1>Delete Page</h1>
    <form method="post">
        <input type="hidden" name="id" value="{{ page.id }}">
        <p>Are you sure you want to delete this page?</p>
        <button type="submit">Delete</button>
    </form>
</body>
</html>
>>>>>>> updated
```

This is a basic HTML structure for the delete page. It includes a form with a hidden input for the page id and a submit button. The form method is set to "post". When the submit button is clicked, the form data will be sent to the server for processing.