aider: Refactored code to search for a template named "home" by default.
# Aider chat conversation:

USER: let's refactor and by default to search for a template named home

ASSISTANT: To refactor the code to search for a template named "home" by default, we need to modify the `loadPage()` method in the `StaticPageController.php` file. We will set the default value of the `$pageName` parameter to "home".

Here is the *SEARCH/REPLACE block* for this change:

```php
src/Controller/StaticPageController.php
<<<<<<< SEARCH
    /**
     * @Route("/page/{pageName}", name="static_page")
     */
    public function loadPage(string $pageName): Response
    {
        $content = $this->themeService->loadStaticPage($pageName);

        return new Response($content);
    }
=======
    /**
     * @Route("/page/{pageName}", name="static_page", defaults={"pageName": "home"})
     */
    public function loadPage(string $pageName = 'home'): Response
    {
        $content = $this->themeService->loadStaticPage($pageName);

        return new Response($content);
    }
>>>>>>> REPLACE
```

This change will make the `loadPage()` method load the "home" page by default when no page name is provided in the URL.