aider: Added `ThemeService` and `StaticPageController` services to the services configuration in `config/services.yaml`.
# Aider chat conversation:

USER: let's also update `config/services.yaml` to add the required services

ASSISTANT: To add the required services, we need to modify the `config/services.yaml` file. We will add the `ThemeService` and `StaticPageController` services to the services configuration.

Here is the *SEARCH/REPLACE block* for the changes:

```yaml
config/services.yaml
<<<<<<< SEARCH
    App\Service\ThemeDiscoveryService:
        public: true

    App\EventSubscriber\CommandSubscriber:
        public: true
        tags:
            - { name: kernel.event_subscriber }

    App\Service\ThemeFilesystemService:
        public: true
        arguments:
            $filesystem: '@Symfony\Component\Filesystem\Filesystem'
            $projectDir: '%kernel.project_dir%'
=======
    App\Service\ThemeDiscoveryService:
        public: true

    App\EventSubscriber\CommandSubscriber:
        public: true
        tags:
            - { name: kernel.event_subscriber }

    App\Service\ThemeFilesystemService:
        public: true
        arguments:
            $filesystem: '@Symfony\Component\Filesystem\Filesystem'
            $projectDir: '%kernel.project_dir%'

    App\Service\ThemeService:
        public: true
        arguments:
            $themesDir: '%kernel.project_dir%/themes'

    App\Controller\StaticPageController:
        public: true
        arguments:
            $themeService: '@App\Service\ThemeService'
>>>>>>> REPLACE
```

This change will add the `ThemeService` and `StaticPageController` services to the services configuration. The `ThemeService` service will be injected with the `themesDir` parameter, which is set to the path of the themes directory. The `StaticPageController` service will be injected with the `ThemeService` service.