aider: Added a new route for the root URL that redirects to the home page.
# Aider chat conversation:

USER: if i access https://localhost/ no page is loaded we should refactor / implement a feature so it will take this into consideration not only routes starting with /page

ASSISTANT: To load the home page by default when accessing the root URL (https://localhost/), we need to modify the `src/Controller/StaticPageController.php` file. We will add a new route for the root URL that redirects to the home page.

Here is the *SEARCH/REPLACE block* for the changes:

```php
src/Controller/StaticPageController.php
<<<<<<< SEARCH
class StaticPageController extends AbstractController
{
    private $themeService;

    public function __construct(ThemeService $themeService)
    {
        $this->themeService = $themeService;
    }

    /**
     * @Route("/page/{pageName}", name="static_page", defaults={"pageName": "home"})
     */
    public function loadPage(string $pageName = 'home'): Response
    {
        $content = $this->themeService->loadStaticPage($pageName);

        return new Response($content);
    }
}
=======
class StaticPageController extends AbstractController
{
    private $themeService;

    public function __construct(ThemeService $themeService)
    {
        $this->themeService = $themeService;
    }

    /**
     * @Route("/", name="home")
     */
    public function home(): Response
    {
        return $this->redirectToRoute('static_page', ['pageName' => 'home']);
    }

    /**
     * @Route("/page/{pageName}", name="static_page", defaults={"pageName": "home"})
     */
    public function loadPage(string $pageName = 'home'): Response
    {
        $content = $this->themeService->loadStaticPage($pageName);

        return new Response($content);
    }
}
>>>>>>> REPLACE
```

This change will add a new route for the root URL that redirects to the home page. When accessing the root URL (https://localhost/), the `home` method will be called, which will redirect to the home page.